It's ulong not int..

+++ DisCatSharp/Entities/Interaction/DiscordInteractionDataOption.cs
@@ -75,7 +75,7 @@ namespace DisCatSharp.Entities
 						ApplicationCommandOptionType.Role => ulong.Parse(this.RawValue),
 						ApplicationCommandOptionType.Mentionable => ulong.Parse(this.RawValue),
 						ApplicationCommandOptionType.Number => double.Parse(this.RawValue),
-						ApplicationCommandOptionType.Attachment => int.Parse(this.RawValue),
+						ApplicationCommandOptionType.Attachment => ulong.Parse(this.RawValue),
 						_ => this.RawValue,
 					};
